Move dask functions outside io.
