--- tests/io/test_asn.py
@@ -1,1 +1,0 @@
-# TODO
